Increase badge text font weight for readability
The `font-weight` property[^1] of badges (rendered on the "Options"
page[^2]) was set to `1`, an unreadably small value. For reference, a
value of `100` is considered `Thin (Hairline)`, and `normal` is `400`.
Changing the value to `normal` makes the text readable again.

The commit the current style was added in, f669979,[^3] does not explain
why such an extremely small value was immediately chosen.

[^1]: https://developer.mozilla.org/en-US/docs/Web/CSS/font-weight
[^2]: https://docs.mitmproxy.org/stable/concepts-options/
[^3]: https://github.com/mitmproxy/mitmproxy/commit/f6699792da82c00de4cdcc1e413bd65caad9d3c3